requirements.txt
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.46
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.1
justwatch==0.7.0
